# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MenuClick API",
    "version": "0.1.0",
    "description": "API do MenuClick — cardápio digital, QR code e delivery.\n\n## Duas superfícies\n\nA API atende dois públicos, e a diferença é o que mais importa ao usá-la:\n\n- **Quem escaneia o QR code** lê o cardápio e faz pedido, sem conta e sem token.\n- **O restaurante** faz todo o resto, com `Authorization: Bearer <token>`.\n\n**Toda rota exige sessão por padrão.** Ser pública é declaração explícita,\ne as poucas que são estão marcadas como tal aqui.\n\n## Convenções\n\n- Listagens respondem um envelope `{ data, limit, offset, total }`, nunca um array cru.\n  `limit` vai de 1 a 100 (default 20); fora da faixa é **400**, não ajuste silencioso.\n- Erro tem sempre a forma `{ statusCode, error, message }`.\n- Valores monetários são **inteiros em centavos**. Nunca float.\n- Nada é apagado: `DELETE` marca o registro como removido, e ele passa a se\n  comportar como se nunca tivesse existido (404 em tudo, fora das listagens).\n- Há limite de requisições por IP: 100/min no geral e **5/min no login**.\n  Estourar responde **429**."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "O token devolvido por `POST /auth/login`. É opaco: não carrega informação, e o servidor guarda só o hash dele."
      }
    },
...
  },
  "paths": {
...
    "/restaurants": {
      "get": {
        "operationId": "listRestaurants",
        "summary": "Os restaurantes da sessão",
        "tags": [
          "Restaurantes"
        ],
        "description": "Devolve o restaurante do usuário logado, não todos: uma listagem geral entregaria o cadastro dos concorrentes. Hoje é sempre um, mas o contrato é de lista.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "in": "query",
            "name": "offset",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "cuisineType": {
                            "type": "string"
                          },
                          "logoUrl": {
                            "type": "string"
                          },
                          "address": {
                            "type": "object",
                            "properties": {
                              "street": {
                                "type": "string",
                                "minLength": 1
                              },
                              "number": {
                                "type": "string",
                                "minLength": 1
                              },
                              "neighborhood": {
                                "type": "string",
                                "minLength": 1
                              },
                              "city": {
                                "type": "string",
                                "minLength": 1
                              },
                              "state": {
                                "type": "string",
                                "minLength": 1
                              },
                              "zipCode": {
                                "type": "string",
                                "minLength": 1
                              }
                            }
                          },
                          "isDelivery": {
                            "type": "boolean"
                          },
                          "isTakeaway": {
                            "type": "boolean"
                          },
                          "isQrcode": {
                            "type": "boolean"
                          },
                          "timezone": {
                            "type": "string"
                          },
                          "acceptingOrders": {
                            "type": "boolean"
                          },
                          "acceptsCash": {
                            "type": "boolean"
                          },
                          "acceptsCardOnDelivery": {
                            "type": "boolean"
                          },
                          "acceptsPix": {
                            "type": "boolean"
                          },
                          "acceptsMealVoucher": {
                            "type": "boolean"
                          },
                          "deliveryFeeMode": {
                            "type": "string"
                          },
                          "deliveryFixedFeeInCents": {
                            "type": "integer"
                          },
                          "freeDeliveryAboveInCents": {
                            "type": "integer"
                          },
                          "deliveryFeeToArrange": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/restaurants/{restaurantId}": {
      "get": {
        "operationId": "getRestaurant",
        "summary": "Detalhe do restaurante",
        "tags": [
          "Restaurantes"
        ],
        "description": "Pedir um restaurante que não é o da sessão responde 404, e não 403: 'proibido' confirmaria que ele existe.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "restaurantId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "cuisineType": {
                      "type": "string"
                    },
                    "logoUrl": {
                      "type": "string"
                    },
                    "address": {
                      "type": "object",
                      "properties": {
                        "street": {
                          "type": "string",
                          "minLength": 1
                        },
                        "number": {
                          "type": "string",
                          "minLength": 1
                        },
                        "neighborhood": {
                          "type": "string",
                          "minLength": 1
                        },
                        "city": {
                          "type": "string",
                          "minLength": 1
                        },
                        "state": {
                          "type": "string",
                          "minLength": 1
                        },
                        "zipCode": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    },
                    "isDelivery": {
                      "type": "boolean"
                    },
                    "isTakeaway": {
                      "type": "boolean"
                    },
                    "isQrcode": {
                      "type": "boolean"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "acceptingOrders": {
                      "type": "boolean"
                    },
                    "acceptsCash": {
                      "type": "boolean"
                    },
                    "acceptsCardOnDelivery": {
                      "type": "boolean"
                    },
                    "acceptsPix": {
                      "type": "boolean"
                    },
                    "acceptsMealVoucher": {
                      "type": "boolean"
                    },
                    "deliveryFeeMode": {
                      "type": "string"
                    },
                    "deliveryFixedFeeInCents": {
                      "type": "integer"
                    },
                    "freeDeliveryAboveInCents": {
                      "type": "integer"
                    },
                    "deliveryFeeToArrange": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateRestaurant",
        "summary": "Edita o restaurante",
        "tags": [
          "Restaurantes"
        ],
        "description": "Edição parcial. O `slug` NÃO é editável por aqui: mudar a URL pública quebraria QR code já impresso, então isso precisa ser operação explícita.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "minProperties": 1,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "cuisineType": {
                    "type": "string",
                    "minLength": 1
                  },
                  "logoUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "address": {
                    "type": "object",
                    "additionalProperties": false,
                    "required": [
                      "street",
                      "number",
                      "neighborhood",
                      "city",
                      "state",
                      "zipCode"
                    ],
                    "properties": {
                      "street": {
                        "type": "string",
                        "minLength": 1
                      },
                      "number": {
                        "type": "string",
                        "minLength": 1
                      },
                      "neighborhood": {
                        "type": "string",
                        "minLength": 1
                      },
                      "city": {
                        "type": "string",
                        "minLength": 1
                      },
                      "state": {
                        "type": "string",
                        "minLength": 1
                      },
                      "zipCode": {
                        "type": "string",
                        "minLength": 1
                      }
                    }
                  },
                  "isDelivery": {
                    "type": "boolean"
                  },
                  "isTakeaway": {
                    "type": "boolean"
                  },
                  "isQrcode": {
                    "type": "boolean"
                  },
                  "timezone": {
                    "type": "string",
                    "pattern": "^[A-Za-z][A-Za-z0-9+_-]*(?:/[A-Za-z0-9+_-]+)*$",
                    "maxLength": 64
                  },
                  "acceptingOrders": {
                    "type": "boolean"
                  },
                  "acceptsCash": {
                    "type": "boolean"
                  },
                  "acceptsCardOnDelivery": {
                    "type": "boolean"
                  },
                  "acceptsPix": {
                    "type": "boolean"
                  },
                  "acceptsMealVoucher": {
                    "type": "boolean"
                  },
                  "deliveryFeeMode": {
                    "type": "string",
                    "enum": [
                      "neighborhood",
                      "fixed"
                    ]
                  },
                  "deliveryFixedFeeInCents": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "freeDeliveryAboveInCents": {
                    "type": "integer",
                    "minimum": 0,
                    "nullable": true
                  },
                  "deliveryFeeToArrange": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "restaurantId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "cuisineType": {
                      "type": "string"
                    },
                    "logoUrl": {
                      "type": "string"
                    },
                    "address": {
                      "type": "object",
                      "properties": {
                        "street": {
                          "type": "string",
                          "minLength": 1
                        },
                        "number": {
                          "type": "string",
                          "minLength": 1
                        },
                        "neighborhood": {
                          "type": "string",
                          "minLength": 1
                        },
                        "city": {
                          "type": "string",
                          "minLength": 1
                        },
                        "state": {
                          "type": "string",
                          "minLength": 1
                        },
                        "zipCode": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    },
                    "isDelivery": {
                      "type": "boolean"
                    },
                    "isTakeaway": {
                      "type": "boolean"
                    },
                    "isQrcode": {
                      "type": "boolean"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "acceptingOrders": {
                      "type": "boolean"
                    },
                    "acceptsCash": {
                      "type": "boolean"
                    },
                    "acceptsCardOnDelivery": {
                      "type": "boolean"
                    },
                    "acceptsPix": {
                      "type": "boolean"
                    },
                    "acceptsMealVoucher": {
                      "type": "boolean"
                    },
                    "deliveryFeeMode": {
                      "type": "string"
                    },
                    "deliveryFixedFeeInCents": {
                      "type": "integer"
                    },
                    "freeDeliveryAboveInCents": {
                      "type": "integer"
                    },
                    "deliveryFeeToArrange": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteRestaurant",
        "summary": "Remove o restaurante",
        "tags": [
          "Restaurantes"
        ],
        "description": "Soft delete: nada sai do banco. O restaurante e os produtos dele passam a se comportar como se nunca tivessem existido (404 em tudo, fora das listagens), na mesma transação. Pedidos NÃO são afetados — eles são histórico, não catálogo.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "restaurantId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Desenvolvimento"
    }
  ],
...
}